refactor: rename files to match convention
