+++ src/matchers/extend-expect.d.ts
@@ -1,7 +1,7 @@
 import type { StyleProp } from 'react-native';
 import type { ReactTestInstance } from 'react-test-renderer';
 import type { TextMatch, TextMatchOptions } from '../matches';
-import type { AccessibilityValueMatcher } from '../helpers/matchers/accessibilityValue';
+import type { AccessibilityValueMatcher } from '../helpers/matchers/match-accessibility-value';
 import type { Style } from './to-have-style';
 
 export interface JestNativeMatchers<R> {
