+++ src/queries/__tests__/hint-text.test.tsx
