+++ src/within.ts
@@ -1,15 +1,15 @@
 import type { ReactTestInstance } from 'react-test-renderer';
 import { bindByTextQueries } from './queries/text';
-import { bindByTestIdQueries } from './queries/testId';
-import { bindByDisplayValueQueries } from './queries/displayValue';
-import { bindByPlaceholderTextQueries } from './queries/placeholderText';
-import { bindByLabelTextQueries } from './queries/labelText';
-import { bindByHintTextQueries } from './queries/hintText';
+import { bindByTestIdQueries } from './queries/test-id';
+import { bindByDisplayValueQueries } from './queries/display-value';
+import { bindByPlaceholderTextQueries } from './queries/placeholder-text';
+import { bindByLabelTextQueries } from './queries/label-text';
+import { bindByHintTextQueries } from './queries/hint-text';
 import { bindByRoleQueries } from './queries/role';
-import { bindByA11yStateQueries } from './queries/a11yState';
-import { bindByA11yValueQueries } from './queries/a11yValue';
-import { bindUnsafeByTypeQueries } from './queries/unsafeType';
-import { bindUnsafeByPropsQueries } from './queries/unsafeProps';
+import { bindByA11yStateQueries } from './queries/accessibility-state';
+import { bindByA11yValueQueries } from './queries/accessibility-value';
+import { bindUnsafeByTypeQueries } from './queries/unsafe-type';
+import { bindUnsafeByPropsQueries } from './queries/unsafe-props';
 
 export function within(instance: ReactTestInstance) {
   return {
